requirements.txt
# FinNet 项目依赖
# 全局依赖文件

# 核心依赖
pydantic-settings>=2.0.0  # 配置管理（common/config.py）
pydantic>=2.0.0

# 爬虫模块依赖
requests>=2.28.0
beautifulsoup4>=4.11.0
chardet>=5.0.0  # 编码检测（SEC HTML 等）
tqdm>=4.64.0
akshare>=1.12.0  # A股数据获取

# 存储层依赖
minio>=7.0.0  # MinIO 对象存储
psycopg2-binary>=2.9.0  # PostgreSQL
sqlalchemy>=2.0.0  # ORM
pymilvus>=2.3.0  # Milvus 向量数据库
neo4j>=5.0.0  # Neo4j 图数据库
elasticsearch>=8.0.0,<9.0.0  # Elasticsearch 全文搜索引擎（兼容 Elasticsearch 8.x 服务器）

# 调度系统
dagster>=1.5.0
dagster-webserver>=1.5.0

# AI/ML 依赖
sentence-transformers>=2.2.0  # Embedding模型（BGE/BCE）

# 其他工具
python-dotenv>=1.0.0  # 环境变量管理
